config: Setup for component refactor
