+++ app/components/Navigation/HeroNavbar.tsx
@@ -9,7 +9,7 @@ import {
 	NavbarMenu,
 	NavbarMenuItem,
 } from '@nextui-org/navbar';
-import NavigationLink from '@/app/components/Navigation/NavigationLink';
+import { NavigationLink } from '@/app/components/navigation/Links';
 import LogoIcon from '@/app/assets/Logo';
 import { Link } from '@nextui-org/link';
 import { Button } from '@nextui-org/button';
@@ -35,9 +35,7 @@ export default function HeroNavbar({ data, theme = 'dark' }: Props) {
 			</NavbarBrand>
 			<NavbarContent className="hidden sm:flex gap-4" justify="center">
 				<NavbarItem>
-					<NavigationLink color="foreground" href="#">
-						Features
-					</NavigationLink>
+					<NavigationLink>Features</NavigationLink>
 				</NavbarItem>
 				<NavbarItem isActive>
 					<NavigationLink href="#" aria-current="page">
@@ -45,14 +43,12 @@ export default function HeroNavbar({ data, theme = 'dark' }: Props) {
 					</NavigationLink>
 				</NavbarItem>
 				<NavbarItem>
-					<NavigationLink color="foreground" href="#">
-						Integrations
-					</NavigationLink>
+					<NavigationLink>Integrations</NavigationLink>
 				</NavbarItem>
 			</NavbarContent>
 			<NavbarContent justify="end">
 				<NavbarItem className="hidden lg:flex">
-					<NavigationLink href="#">Login</NavigationLink>
+					<NavigationLink>Login</NavigationLink>
 				</NavbarItem>
 				<NavbarItem>
 					<Button as={Link} color="primary" href="#" variant="flat">
